requirements.txt
plotly==4.14.3
dash==1.20.0
pandas==1.2.3
requests==2.24.0
tethysts==0.2.5
rtree==0.9.7
pyyaml==5.3.1
flask-caching==1.10.1
nbformat==5.1.3
